handle comment window within list
